Converted the bit depths of IV data from u32 depth to usize depth

+++ core/src/bpcs/capacity.rs
@@ -8,18 +8,18 @@ use crate::{
 
 pub(crate) fn calculate_num_of_embedding_planes(
     min_alpha: f64,
-    message_plane_length: u32,
+    message_plane_length: usize,
 ) -> usize {
     let iv_plane_num = calculate_iv_plane_number(min_alpha);
     let conj_map_plane_num =
-        num_of_prefixed_planes_for_n_bits(message_plane_length as usize, prefix_length(min_alpha));
+        num_of_prefixed_planes_for_n_bits(message_plane_length, prefix_length(min_alpha));
 
     iv_plane_num + conj_map_plane_num + message_plane_length as usize
 }
 
 pub(crate) fn check_capacity(
     min_alpha: f64,
-    message_plane_length: u32,
+    message_plane_length: usize,
     accepted_planes_num: usize,
 ) -> Result<(), SteganographyError> {
     let required_min_accepted_num =
